+++ checks/AZS/AZS008_registration_entries_sorted/AZS008_test.go
@@ -1,0 +1,18 @@
+package AZS008
+
+import (
+	"path/filepath"
+	"runtime"
+	"testing"
+
+	"golang.org/x/tools/go/analysis/analysistest"
+)
+
+func TestAZS008(t *testing.T) {
+	t.Parallel()
+
+	_, filename, _, _ := runtime.Caller(0)
+	dir := filepath.Join(filepath.Dir(filename), "testdata")
+
+	analysistest.RunWithSuggestedFixes(t, dir, Analyzer, "azs008")
+}